Big and important refactoring!
