--- cmd/command/common.go
@@ -1,5 +1,0 @@
-package command
-
-const (
-	versionPackage = "./cmd/version"
-)
