+++ cmd/v1/cli/cli_test.go
@@ -1,0 +1,1 @@
+package cli
