+++ internal/service/semver/test/package-invalid-version.json
