+++ internal/service/util/util.go
