+++ internal/v1/spec/test/min.yaml
@@ -1,4 +1,4 @@
-version: "1"
+version: "1.0"
 
 language: go
