+++ pkg/cli/cli_test.go
@@ -1,4 +1,4 @@
-package command
+package cli
 
 import (
 	"testing"
